Accessor: Move to `lofty::tag`
