+++ src/id3/v1/tag.rs
@@ -1,8 +1,8 @@
 use crate::config::WriteOptions;
 use crate::error::{LoftyError, Result};
 use crate::id3::v1::constants::GENRES;
-use crate::tag::{ItemKey, ItemValue, Tag, TagExt, TagItem, TagType};
-use crate::traits::{Accessor, MergeTag, SplitTag};
+use crate::tag::{Accessor, ItemKey, ItemValue, Tag, TagExt, TagItem, TagType};
+use crate::traits::{MergeTag, SplitTag};
 
 use std::borrow::Cow;
 use std::fs::File;
