+++ src/iff/aiff/tag.rs
@@ -2,8 +2,8 @@ use crate::config::WriteOptions;
 use crate::error::{LoftyError, Result};
 use crate::iff::chunk::Chunks;
 use crate::macros::err;
-use crate::tag::{ItemKey, ItemValue, Tag, TagExt, TagItem, TagType};
-use crate::traits::{Accessor, MergeTag, SplitTag};
+use crate::tag::{Accessor, ItemKey, ItemValue, Tag, TagExt, TagItem, TagType};
+use crate::traits::{MergeTag, SplitTag};
 
 use std::borrow::Cow;
 use std::fs::File;
